backend/requirements.txt
fastapi[standard]
pydantic
aiofiles
aiohttp
boto3
langchain-core
logger
requests
numpy
uvicorn
pymongo[aws]
python-jose[cryptography]
python-docx
python-dotenv
bcrypt
email-validator
reportlab
jinja2
sse-starlette
openpyxl
pandas
httpx
celery[redis]
dirtyjson